update validation when delete show/challenge

+++ app/scripts/app/shows/list/controller.js
@@ -36,26 +36,30 @@
     };
 
     $scope.deleteShow = function(item, index) {
-      $uibModal.open({
-        controller: 'ConfirmModalController',
-        controllerAs: 'cm',
-        templateUrl: 'scripts/components/confirmModal/view.html',
-        resolve: {
-          title: function() {
-            return 'Delete Show Confirmation';
-          },
-          description: function() {
-            return 'Are you sure that you want to delete this show?';
+      if (item.status === 'unpublished') {
+        $uibModal.open({
+          controller: 'ConfirmModalController',
+          controllerAs: 'cm',
+          templateUrl: 'scripts/components/confirmModal/view.html',
+          resolve: {
+            title: function() {
+              return 'Delete Show Confirmation';
+            },
+            description: function() {
+              return 'Are you sure that you want to delete this show? It will remove all challenges belong to this.';
+            }
           }
-        }
-      }).result.then(function() {
-        ShowService.delete({id: item._id}).$promise.then(function() {
-          $scope.items.splice(index, 1);
-          $scope.total--;
-        }).catch(function() {
-          growl.error('Error when delete a show. Please try again');
+        }).result.then(function() {
+          ShowService.delete({id: item._id}).$promise.then(function() {
+            $scope.items.splice(index, 1);
+            $scope.total--;
+          }).catch(function() {
+            growl.error('Error when delete a show. Please try again');
+          });
         });
-      });
+      } else {
+        growl.error('Cannot delete unpublished show');
+      }
     };
   }
 })();
